Day 32 and 33 - Self-referencing models and using query strings.

+++ dpproject/dpproject/views.py
@@ -10,8 +10,11 @@ def AddPage(request,x,y):
     return HttpResponse(x + y);
     
 def HomePage(request):
-    return redirect('/accounts/login')
-    
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
+    return redirect('/todo')
+
 def AboutPage(request):
     msg = ""
     parms = {"msg": msg}
@@ -24,6 +27,9 @@ def Looper(request,value,loop):
     return HttpResponse("Success")
 
 def completeTask(request,id):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     complete = lkpStatus.objects.get(pk=3)
     print(complete)
     task = Task.objects.get(pk=id)
@@ -47,6 +53,9 @@ def userRegister(request):
     return render(request,'register.html',params)
 
 def CreateTask(request):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     data = Task()
     data.owner = request.user
     form = forms.TaskForm(instance=data)
@@ -63,6 +72,9 @@ def CreateTask(request):
     return render(request,'TaskEdit.html',parms)
 
 def deleteTasks(request):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     Task.objects.all().delete()
     tasks = Task.objects.all()
     title = "All Tasks"
@@ -87,6 +99,9 @@ def test(request):
     return render(request,'TaskList.html',parms)
 
 def massUpdate(request):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     Task.objects.filter(minutes_spent == 0 ).update(minutes_spent = 2)
     tasks = Task.objects.all()
     title = "All Tasks"
@@ -94,9 +109,13 @@ def massUpdate(request):
     return render(request,'TaskList.html',parms)
 
 def EditTask(request,id):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     data = Task.objects.get(pk=id)
     data.owner = request.user
     form = forms.TaskForm(instance=data)
+    parent = []
     msg = "Please update data for the record"
     if request.method == "POST":
         if 'delete' in request.POST:
@@ -111,23 +130,39 @@ def EditTask(request,id):
                 return redirect('/tasks')
             else:
                 msg = "Invalid Record"
+        if request.POST['tf']:
+            pid = int(request.POST['tf'])
+            parent = Timeframe.objects.get(pk=pid)
+    else:
+        if request.GET['tf']:
+            pid = int(request.GET['tf'])
+            parent = Timeframe.objects.get(pk=pid)
     title = "Edit Task"
-    parms = {"form": form, "title": title, "msg": msg}
+    parms = {"form": form, "title": title, "msg": msg, "parent": parent}
     return render(request,'TaskEdit.html',parms)
 
 def ListTasks(request):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     tasks = Task.objects.all().order_by('duedate','name')
     title = "All Tasks"
     parms = {"title": title, "tasks": tasks}
     return render(request,'TaskList.html',parms)
 
 def ListRoadmaps(request):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     items = Roadmap.objects.all().order_by('name')
     title = "My Roadmaps"
     parms = {"title": title, "items": items}
     return render(request,'RoadmapList.html',parms)
 
 def CreateRoadmap(request):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     data = Roadmap()
     data.owner = request.user
     form = forms.RoadmapForm(instance=data)
@@ -146,6 +181,9 @@ def CreateRoadmap(request):
     return render(request,'RoadmapEdit.html',parms)
 
 def EditRoadmap(request,id):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     data = Roadmap.objects.get(pk=id)
     data.owner = request.user
     form = forms.RoadmapForm(instance=data)
@@ -166,10 +204,13 @@ def EditRoadmap(request,id):
                 msg = "Invalid Record"
     title = "Edit Roadmap"
     tfs = Timeframe.objects.filter(roadmap=data).order_by('name')
-    parms = {"form": form, "title": title, "msg": msg, "timeframes": tfs}
+    parms = {"form": form, "title": title, "msg": msg, "timeframes": tfs, "rmid": id}
     return render(request,'RoadmapEdit.html',parms)
 
 def ActiveTasks(request):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     cancel = lkpStatus.objects.get(shortname='CNCL')
     complete = lkpStatus.objects.get(shortname='COMP')
     tasks = Task.objects.filter(~Q(status=complete)).order_by('duedate','name')
@@ -204,8 +245,12 @@ def showname(request):
     return HttpResponse(name)
 
 def EditTimeframe(request,id):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     data = Timeframe.objects.get(pk=id)
     form = forms.TimeframeForm(instance=data)
+    parent = []
     msg = "Please update data for the record"
     if request.method == "POST":
         if 'delete' in request.POST:
@@ -220,19 +265,33 @@ def EditTimeframe(request,id):
                 return redirect('/tfs')
             else:
                 msg = "Invalid Record"
+        if request.POST['rm']:
+            pid = int(request.POST['rm'])
+            parent = Roadmap.objects.get(pk=pid)
+    else:
+        if request.GET['rm']:
+            pid = int(request.GET['rm'])
+            parent = Roadmap.objects.get(pk=pid)
+
     title = "Edit Time Frame"
     tasks = Task.objects.filter(timeframe=data).order_by('duedate','name')
 
-    parms = {"form": form, "title": title, "msg": msg, "tasks": tasks}
+    parms = {"form": form, "title": title, "msg": msg, "tasks": tasks, "tfid": id, "parent": parent}
     return render(request,'TimeframeEdit.html',parms)
 
 def ListTimeframes(request):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     tfs = Timeframe.objects.all().order_by('name')
     title = "All Time Frames"
     parms = {"title": title, "timeframes": tfs}
     return render(request,'TimeframeList.html',parms)
 
 def CreateTimeframe(request):
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
+
     data = Timeframe()
     form = forms.TimeframeForm(instance=data)
     msg = "Create a New Record"
